Clear code, add loading view, edit fragment. adapter

+++ app/src/main/java/com/example/reddittest/db/PostEntity.kt
@@ -11,7 +11,7 @@ data class PostEntity(
     val authorName: String?,
     val title: String?,
     val body: String?,
-    val date: Double?,
+    val date: Long?,
     val imageUrl: String? = "",
     val commentsCount: Int?
 ) {
